Renaming variables. Remove xml approach. optimize code-

+++ src/main/java/de/reza/documentclassifier/utils/DatasetProcessor.java
@@ -1,5 +1,6 @@
 package de.reza.documentclassifier.utils;
 
+import lombok.extern.slf4j.Slf4j;
 import net.lingala.zip4j.ZipFile;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -10,6 +11,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 
 @Service
+@Slf4j
 public class DatasetProcessor {
 
     /**
@@ -21,16 +23,22 @@ public class DatasetProcessor {
     public String unzip(MultipartFile dataset, String uuid){
 
         File dic = new File("models/" + uuid);
-        dic.mkdirs();
-        Path filepath = Paths.get("models/" + uuid, dataset.getOriginalFilename());
-        try {
-            dataset.transferTo(filepath);
-            new ZipFile(filepath.toFile()).extractAll("models/" + uuid + "/dataset/");
-            filepath.toFile().delete();
-            return "models/" + uuid + "/dataset/";
-        } catch (IOException e) {
-            throw new RuntimeException(e);
+        boolean createFolderSuccessful = dic.mkdirs();
+        if(createFolderSuccessful) {
+            Path filepath = Paths.get("models/" + uuid, dataset.getOriginalFilename());
+            try {
+                dataset.transferTo(filepath);
+                new ZipFile(filepath.toFile()).extractAll("models/" + uuid + "/trainingFiles/");
+                boolean removeZipFile = filepath.toFile().delete();
+                if(removeZipFile){
+                    log.info("Zip file deleted");
+                    return "models/" + uuid + "/trainingFiles/";
+                }
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
         }
+        return null;
     }
 
 }
